separate logic from graphic
