+++ ts/components/messages/paginated/MessageList/index.tsx
@@ -12,7 +12,7 @@ import {
 import { connect } from "react-redux";
 
 import { maximumItemsFromAPI, pageSize } from "../../../../config";
-import { UaDonationBanner } from "../../../../features/uaDonations/components/UaDonationBanner";
+import { UaDonationsBanner } from "../../../../features/uaDonations/components/UaDonationsBanner";
 import I18n from "../../../../i18n";
 import {
   loadNextPageMessages,
@@ -246,7 +246,7 @@ const MessageList = ({
       {isRefreshing && isFirstLoad && <Loader />}
 
       <AnimatedFlatList
-        ListHeaderComponent={<UaDonationBanner />}
+        ListHeaderComponent={<UaDonationsBanner />}
         ItemSeparatorComponent={ItemSeparator}
         ListEmptyComponent={renderEmptyList({
           error,
